requirements.txt
click==8.0.1
colorama==0.4.4
Flask==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
python-dotenv
requests
gunicorn==19.3.0
firebase
firebase_admin
selenium
Pillow
pyrebase
pyautogui
